+++ module/actes-generique/action/IParapheurEnvoie.class.php
@@ -1,54 +1,9 @@
 <?php
 
+require_once __DIR__."/../../../connecteur-type/signature/SignatureEnvoie.class.php";
 
-class IParapheurEnvoie extends ActionExecutor {
-
-	/**
-	 * @return bool
-	 * @throws Exception
-	 */
-	public function go(){
-
-        /** @var IParapheur $signature */
-		$signature = $this->getConnecteur('signature');
-		
-		$actes = $this->getDonneesFormulaire();
-		
-		$file_content = file_get_contents($actes->getFilePath('arrete'));
-		$finfo = new finfo(FILEINFO_MIME);
-		$content_type = $finfo->file($actes->getFilePath('arrete'),FILEINFO_MIME_TYPE);
-
-		$annexe = array();
-		if ($actes->get('autre_document_attache')) {
-			foreach($actes->get('autre_document_attache') as $num => $fileName ){
-				$annexe_content =  file_get_contents($actes->getFilePath('autre_document_attache',$num));
-				$annexe_content_type = $finfo->file($actes->getFilePath('autre_document_attache',$num),FILEINFO_MIME_TYPE);
-					
-				$annexe[] = array(
-					'name' => $fileName,
-					'file_content' => $annexe_content,
-					'content_type' => $annexe_content_type,
-				);
-				
-			}
-		}
-		$signature->setSendingMetadata($actes);
-		$dossierID = $signature->getDossierID($actes->get('numero_de_lacte'),$actes->get('objet'));
-		$result = $signature->sendDocument($actes->get('iparapheur_type'),
-											$actes->get('iparapheur_sous_type'),
-											$dossierID,
-											$file_content,
-											$content_type,
-                                            $annexe);
-		if (! $result){
-			$this->setLastMessage("La connexion avec le iParapheur a échoué : " . $signature->getLastError());
-			return false;
-		}
-		
-		$this->addActionOK("Le document a été envoyé au parapheur électronique");
-		$this->notify($this->action, $this->type,"Le document a été envoyé au parapheur électronique");
-		
-		return true;
-	}
-	
-}
+/**
+ * Class IParapheurEnvoie
+ * @deprecated PA 2.1
+ */
+class IParapheurEnvoie extends SignatureEnvoie { /* Nothing to do*/ }